# Conflicts:
#	yukon-client/Servers/Web Application Server.launch
#	yukon-client/common/.classpath

+++ yukon-client/common/src/com/cannontech/common/util/SftpConnection.java
@@ -33,7 +33,6 @@ public class SftpConnection implements AutoCloseable {
     private final String username;
     private final String password;
     private File privateKeyFile;
-    private File publicKeyFile;
     
     public SftpConnection(String domain, String port, Optional<YukonHttpProxy> proxy, String username, String password, 
                        String privateKey) throws IOException {
@@ -98,11 +97,12 @@ public class SftpConnection implements AutoCloseable {
     
     @Override
     public void close() {
-        try {
-            Files.delete(publicKeyFile.toPath());
-            Files.delete(privateKeyFile.toPath());
-        } catch (Exception e) {
-            log.debug("Error deleting key file", e);
+        if (privateKeyFile != null) {
+            try {
+                Files.delete(privateKeyFile.toPath());
+            } catch (Exception e) {
+                log.debug("Error deleting key file", e);
+            }
         }
     }
